Changes in method generator classes

+++ Dynamic.Tekla.Structures.API/Dynamic.Tekla.Structures/Tekla.Structures.Drawing/AngleDimension.cs
@@ -160,7 +160,7 @@ public System.Double GetAngle()
 
     try
     {
-        	var result = teklaObject.GetAngle();
+        	var result = (System.Double) teklaObject.GetAngle();
 
         	return result;
     }
